+++ pages/post/create/index.js
@@ -1,13 +1,11 @@
 import CreatePostFirst from './CreatePostFirst';
 import CreatePostSecond from './CreatePostSecond';
 import CreatePostThird from './CreatePostThird';
-import CreatePostFourth from './CreatePostFourth';
 import SuccessPost from './SuccessPost';
 
 module.exports = {
   CreatePostFirst,
   CreatePostSecond,
   CreatePostThird,
-  CreatePostFourth,
   SuccessPost,
 };
